Add a typechecker for queries
This includes an overhaul of the query IR that enables reuse of the same IR in
the pre- and post-typechecked forms.  It also introduces a data-driven
specification of the library API (i.e., the type providers for CodeQL that are
supported).  The library is specified in a KDL data file that includes typed
methods for each object that can be queried.

The library definition is currently used to drive the type checker (i.e., to
determine the return type of methods).

In the future, it will also be used to generate documentation.