Refactor planning of the planner to be more data driven
There is now a map of methods to their planner implementations. This simplifies
the code and enables the handlers to be defined separately more easily. It also
enables (relatively) principled cross-checking to ensure that the actual
implementations at run-time reflect what the library promises (i.e., if we claim
a method is implemented, it really is).

This change also refactors the storage of some global maps (related to the
library) to use `once_cell` to make them globals initialized once (rather than
scoped more modestly as locals like they were before).