Make the query parser more robust
There were a number of assumptions in the query parser (and there are still a
few) that use explicit child node indexing.  This is unsafe, as comments can
appear almost anywhere and shift the indexing around.

This commit changes the parser to query child nodes by name and kind where
appropriate.  There are still a few indexing operations, but they seem a bit
safer (e.g., for qualified expressions). That isn't a guarantee that it is safe
and we'll need further testing.