Revise the evaluation strategy
The plan IR is now much simpler, and so is the evaluator. The planner does more
work to basically CPS the entire computation and (mostly) fuse all operations
into single lambdas.

There is now just one constructor in the plan IR (and the evaluation IR) for
each type that can be returned by a node filter.