Implement evaluation of relational values
Relational values were supported in aggregate contexts before where they were
immediately eliminated.  This change extends the planner and evaluator to be
able to evaluate them more generally. This change comes with support for a new
method: `Parameter.getName`. While this is not itself relational, it was
impossible to use previously since the only way to get a parameter was
`Method.getAParameter`. The combination of the two is now supported.

This change required two major extensions:

- The type checker now allows relational values in many more contexts
- The planner supports evaluation of relational values as lists in some contexts

All of this ended up requiring a refactoring that I was expecting. Many values
that were previously boxed are now ref counted, as we are now constructing
closures with strange lifetimes that can't just take ownership of their inputs.

This change also includes a major fix to the integration tests. Before, errors
weren't being detected because it was cleanly returning `Err`. Now it fails
loudly with `unwrap`, which revealed that all of the tests were actually
broken (for a silly reason).