Improve the API of the filter transformer function
This function lifts operations from scalars to lists of values during
evaluation.  Before, it always returned a transformer (unless it panicked).
This meant that the caller had to test to see if the argument was a supported
type (i.e., a list).

This was cumbersome, as it meant that as additional list types were added, they
had to be checked for in multiple places.  Instead, return an Option where None
indicates that the argument type is unsupported.  This way, callers just need to
try and rely on the transformer function itself to report what is possible.