Extend the type system and library
The new library methods and types are not yet implemented.  This change adds the
necessary type definitions, including some new structure in the library.kdl file
that reflects the type containment structure of the CodeQL types.  This also
includes a tweak to the library indexing structure to include the new
information.

This is being added to further support enhanced query planning that will be
required to implement the call/argument library methods.