Add support for binding relational values to variables
The construct in question is shown in the new example:

```
from Method m, Parameter p
where p = m.getAParameter() and p.getName() = "log"
select m
```

The relational value `m.getAParameter()` is bound to `p` so that it can be
referenced multiple times.  It compiles down to a loop where `p` is bound to
each value produced by the relational value assigned to it.

The implementation introduces a new preprocessing pass, query planning, before
query compilation. The query planner is a general rewriter, but currently only
handles the rewrite of terms like the above into a new `Binder` form that is
evaluated as described.

Limitations:

- Only relational parameter expressions are currently supported
- The rewriter does not reorder any terms, so flipping the order of the terms in
the example will break