"Improve" parallelism
Before, `ql-grep` was hard-coded to use 8 threads to search in parallel.  Now
there is a command line option to control the number of threads to use, with the
default being all threads supported by the CPU.

This is very greedy and unkind to everything else on the system, but is very
fast.  The heuristic could change in the future.