Add a feature to dump the parsed/transformed query IRs
This option prints all of the printable IRs before exiting (and not running the
query).  This is a debugging tool to understand how parsing and query planning
actually see the query.

The compiled query can't easily be printed because it is just a web of nested
functions.  That should not be a major problem in general, however, as
compilation is a very direct translation of the planned query IR.