Add support for queries that can refer to file imports/includes
There are three primary new methods:

- `Callable.getFile`
- `File.getAnImport`
- `Import.getName`

Together, these enable the user to query the list of imports (or includes) in
the file and use them in predicates.  This required new infrastructure to
compute file-level indexes (of imports/includes) before evaluating queries.  The
new indexing pass is driven by tags in the library that indicate which
preprocessing pass each method that requires one depends on.  The system will
automatically run and make the pass available if (and only if) it is required to
evaluate the query.