Highlight syntax elements that contribute to query matches
We now track the source ranges of syntax elements that are referenced by every
sub-expression in queries. We keep those that contribute to successful matches
and discard those that do not. The source ranges are saved and passed to the
result printer, which now highlights them in bold using ANSI escape sequences.

The major mechanism is a change to the `NodeMatcher` type (and new
`NodeListMatcher` type) to wrap results with a set of source ranges to track.

This commit also includes a more modular mechanism for printing matches,
including the new ANSI escape sequence support.