Add new predicates for querying the call sites in functions/methods/callables
The new predicate is `getACall`, which returns a relation containing all call
sites in the associated function/method/callable. That is complemented with a
new type, `Call`, which has a new method `getTarget`.  For now, the target is
just a string that can be matched against.  See the test cases for example
queries.