Enable query evaluation functions to fail gracefully
This changes the return value for scalar evaluators into an `Option` so that
evaluations that are impossible can fail gracefully.  This is motivated by the
new callsite argument indexing function (`Call.getArgument`), which takes an
integer index.  If the index is out of bounds for a particular call site, it
must be able to fail gracefully (and without aborting evaluation, as that may be
fine).